Implement float and double static variable initialization for BESM-6.
- Emit ,real, directive for TAC_STATIC_INIT_FLOAT and TAC_STATIC_INIT_DOUBLE
- Use %.13g format to cover BESM-6's 39-bit mantissa precision
- Promote Tac_Const.u.float_val and Tac_StaticInit.u.float_val from float to
  double so no precision is lost before the Madlen assembler converts the
  decimal literal to BESM-6 real format (IEEE float32 has less precision than
  BESM-6 real)
- Remove the (float) truncation casts in const_convert, tac_import, tac_export

+++ semantic/const_convert.c
@@ -191,7 +191,7 @@ Tac_StaticInit *new_static_init_from_literal(const Type *target_type, const Lite
 
     case TYPE_FLOAT:
         result              = tac_new_static_init(TAC_STATIC_INIT_FLOAT);
-        result->u.float_val = (float)literal_to_double(lit);
+        result->u.float_val = literal_to_double(lit);
         break;
 
     case TYPE_DOUBLE:
